requirements.txt
PyGithub==1.35
GitPython==2.1.8
bandit==1.4.0